requirements.txt
pandas>=2.0.0
tqdm>=4.0.0
apache-airflow==2.8.1 